NOTICK:  rename cpi-info projects for consistency
Renamed to match some of the other projects (config etc.), and also split out impl into separate projects.

Simplified and removed superfluous interfaces.
