+++ applications/experimental/virtual-node-cli/build.gradle
@@ -21,7 +21,7 @@ dependencies {
     implementation project(":libs:virtual-node:virtual-node-info")
     implementation project(":components:virtual-node:virtual-node-common")
     implementation project(":components:virtual-node:virtual-node-info-service-writer")
-    implementation project(":components:virtual-node:cpi-info-service-writer")
+    api project(":components:virtual-node:cpi-info-write-service")
     implementation project(':components:kafka-topic-admin')
     implementation project(":applications:tools:common-setup")
     implementation project(':components:configuration:configuration-publish-service')
@@ -38,7 +38,7 @@ dependencies {
     implementation project(":libs:configuration:configuration-core")
 
     implementation project(":components:virtual-node:virtual-node-info-service-writer")
-    implementation project(":components:virtual-node:cpi-info-service-writer")
+    implementation project(":components:virtual-node:cpi-info-write-service")
 
     runtimeOnly project(":components:configuration:configuration-read-service")
 
@@ -50,6 +50,7 @@ dependencies {
     runtimeOnly project(":libs:kafka-utils:topic-admin-impl")
     runtimeOnly project(":libs:schema-registry:schema-registry-impl")
     runtimeOnly project(":libs:lifecycle:lifecycle-impl")
+    runtimeOnly project(":components:virtual-node:cpi-info-write-service-impl")
 
     runtimeOnly "org.apache.logging.log4j:log4j-slf4j-impl:$log4jVersion"
     runtimeOnly "org.apache.felix:org.apache.felix.scr:$felixScrVersion"
