+++ components/virtual-node/cpi-info-read-service-impl/src/test/kotlin/net/corda/cpiinfo/read/CpiInfoMapTest.kt
