+++ components/virtual-node/cpi-info-read-service/readme.md
@@ -1,0 +1,5 @@
+# CPI Info Service Implementation
+
+For a given CPI identifier, return the associated meta-data.
+
+The meta-data can be used to then used to locate CPKs and other tasks.
