+++ components/virtual-node/cpi-info-read-service/src/main/kotlin/net/corda/cpiinfo/read/CpiInfoListener.kt
@@ -1,0 +1,10 @@
+package net.corda.cpiinfo.read
+
+import net.corda.packaging.CPI
+
+/**
+ * Functional interface that notifies us of any changes to a [CPI.Identifier] and its [CPI.Metadata]
+ */
+fun interface CpiInfoListener {
+    fun onUpdate(changedKeys: Set<CPI.Identifier>, currentSnapshot: Map<CPI.Identifier, CPI.Metadata>)
+}
