--- components/virtual-node/cpi-info-service-reader/src/main/kotlin/net/corda/cpiinfo/read/CpiInfoListener.kt
@@ -1,7 +1,0 @@
-package net.corda.cpiinfo.read
-
-import net.corda.packaging.CPI
-
-fun interface CpiInfoListener {
-    fun onUpdate(changedKeys: Set<CPI.Identifier>, currentSnapshot: Map<CPI.Identifier, CPI.Metadata>)
-}
