--- components/virtual-node/cpi-info-service-reader/src/main/kotlin/net/corda/cpiinfo/read/CpiInfoReader.kt
@@ -1,15 +1,0 @@
-package net.corda.cpiinfo.read
-
-import net.corda.packaging.CPI
-
-/**
- * CPI Info service.  Returns CPI metadata for a given identifier.
- */
-interface CpiInfoReader {
-    /**
-     * Returns a CPI metadata for a given identifier, or `null` if not found.
-     */
-    fun get(identifier: CPI.Identifier): CPI.Metadata?
-
-    fun registerCallback(listener: CpiInfoListener): AutoCloseable
-}
