--- components/virtual-node/cpi-info-service-reader/src/main/kotlin/net/corda/cpiinfo/read/CpiInfoReaderComponent.kt
@@ -1,5 +1,0 @@
-package net.corda.cpiinfo.read
-
-import net.corda.lifecycle.Lifecycle
-
-interface CpiInfoReaderComponent : CpiInfoReader, Lifecycle
