--- components/virtual-node/cpi-info-service-writer/src/main/kotlin/net/corda/cpiinfo/write/CpiInfoWriterComponent.kt
@@ -1,5 +1,0 @@
-package net.corda.cpiinfo.write
-
-import net.corda.lifecycle.Lifecycle
-
-interface CpiInfoWriterComponent : CpiInfoWriter, Lifecycle
