+++ components/virtual-node/cpi-info-write-service/build.gradle
@@ -1,0 +1,43 @@
+plugins {
+    id 'com.r3.internal.gradle.plugins.r3Publish'
+    id 'corda.common-library'
+}
+
+description "CPI Info Service Writer Interface"
+
+dependencies {
+    compileOnly 'org.osgi:osgi.annotation'
+    compileOnly "org.osgi:org.osgi.service.component.annotations"
+
+    implementation platform("net.corda:corda-api:$cordaApiVersion")
+    implementation "net.corda:corda-packaging"
+    implementation "net.corda:corda-base"
+    implementation "net.corda:corda-avro-schema"
+    implementation "net.corda:corda-topic-schema"
+    implementation "net.corda:corda-packaging-avro-converters"
+
+    implementation project(":libs:configuration:configuration-core")
+    implementation project(":libs:lifecycle:lifecycle")
+    implementation project(":libs:messaging:messaging")
+
+    implementation "com.typesafe:config:$typeSafeConfigVersion"
+
+    implementation project(":libs:virtual-node:virtual-node-info")
+    implementation project(":components:virtual-node:virtual-node-common")
+    implementation project(":components:virtual-node:cpi-info-read-service")
+
+    implementation project(":libs:configuration:configuration-core")
+    implementation project(":libs:configuration:configuration-read")
+    implementation project(":components:configuration:configuration-read-service")
+
+
+    runtimeOnly project(":libs:configuration:configuration-read-impl")
+    runtimeOnly project(":components:configuration:configuration-read-service-impl")
+
+    testImplementation project(":libs:lifecycle:lifecycle-impl")
+    testImplementation project(":libs:lifecycle:registry")
+
+    testImplementation "org.assertj:assertj-core:$assertjVersion"
+    testImplementation "org.mockito:mockito-core:$mockitoVersion"
+    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
+}
