+++ components/virtual-node/cpi-info-write-service/detekt-baseline.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" ?>
+<SmellBaseline>
+  <ManuallySuppressedIssues></ManuallySuppressedIssues>
+  <CurrentIssues></CurrentIssues>
+</SmellBaseline>
