+++ components/virtual-node/cpi-info-write-service/readme.md
@@ -1,0 +1,5 @@
+# CPI Info Service Writer Interface
+
+For a given CPI identifier, put the meta data on Kafka.
+
+The meta-data can be used to then used to locate CPKs and other tasks.
