+++ settings.gradle
@@ -158,8 +158,10 @@ include 'components:sandbox:sandbox-service:sandbox-service-test-resources:sandb
 include 'components:sandbox:sandbox-service:sandbox-service-test-resources:hello-world-cpk-one'
 include 'components:sandbox:sandbox-service:sandbox-service-test-resources:hello-world-cpk-two'
 include 'components:sandbox:sandbox-service:sandbox-service-test-resources:hello-world-cpk-three'
-include 'components:virtual-node:cpi-info-service-reader'
-include 'components:virtual-node:cpi-info-service-writer'
+include 'components:virtual-node:cpi-info-read-service'
+include 'components:virtual-node:cpi-info-read-service-impl'
+include 'components:virtual-node:cpi-info-write-service'
+include 'components:virtual-node:cpi-info-write-service-impl'
 include 'components:virtual-node:sandbox-group-context-service'
 include 'components:virtual-node:virtual-node-common'
 include 'components:virtual-node:virtual-node-info-service-reader'
